Keep billing guards outside the route
